pyproject.toml
[project]
name = "ipgiverbot"
version = "0.1.0"
description = "send ip in telegrambot"
readme = "README.md"
requires-python = ">=3"
dependencies = [
    "pytelegrambotapi>=4.29",
    "python-dotenv>=1.2",
]
